requirements.txt
openai~=0.27.8
vocode~=0.1.110
fastapi~=0.95.2
python-dotenv~=1.0.0
requests~=2.31.0
beautifulsoup4~=4.12.2
pyngrok~=6.0.0
matplotlib
plotly
sklearn